Added command handlers for shipment info

+++ src/EA.Iws.Web/ViewModels/Shipment/ShipmentInfoViewModel.cs
@@ -15,7 +15,7 @@
         [Range(1, 99999, ErrorMessage = "The number of shipments must be at least 1 and cannot be greater than 99999")]
         public int NumberOfShipments { get; set; }
 
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         public ShipmentQuantityUnits Units { get; set; }
